Some WebRtcIsac_* and WebRtcIsacfix_* functions: type encoded stream as uint8[]
The affected functions are

  WebRtcIsacfix_ReadFrameLen
  WebRtcIsacfix_GetNewBitStream
  WebRtcIsacfix_ReadBwIndex

and

  WebRtcIsac_ReadFrameLen
  WebRtcIsac_GetNewBitStream
  WebRtcIsac_ReadBwIndex
  WebRtcIsac_GetRedPayload

BUG=909
R=aluebs@webrtc.org, henrik.lundin@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/22979004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7429 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/modules/audio_coding/codecs/isac/fix/interface/isacfix.h
@@ -357,7 +357,7 @@ extern "C" {
    *
    */
 
-  int16_t WebRtcIsacfix_ReadFrameLen(const int16_t* encoded,
+  int16_t WebRtcIsacfix_ReadFrameLen(const uint8_t* encoded,
                                      int encoded_len_bytes,
                                      int16_t* frameLength);
 
@@ -557,7 +557,7 @@ extern "C" {
   int16_t WebRtcIsacfix_GetNewBitStream(ISACFIX_MainStruct *ISAC_main_inst,
                                         int16_t          bweIndex,
                                         float              scale,
-                                        int16_t        *encoded);
+                                        uint8_t* encoded);
 
 
   /****************************************************************************
@@ -608,7 +608,7 @@ extern "C" {
    *
    */
 
-  int16_t WebRtcIsacfix_ReadBwIndex(const int16_t* encoded,
+  int16_t WebRtcIsacfix_ReadBwIndex(const uint8_t* encoded,
                                     int encoded_len_bytes,
                                     int16_t* rateIndex);
